pyproject.toml
[project]
name = "storylens"
version = "0.1.0"
description = "storylens using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai-tools>=1.12.2",
    "crewai[tools]==1.12.2",
    "fpdf2>=2.8.7",
    "gradio>=6.10.0",
    "openai>=2.30.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sendgrid>=6.12.5",
]

[project.scripts]
storylens = "storylens.main:run"
run_crew = "storylens.main:run"
train = "storylens.main:train"
replay = "storylens.main:replay"
test = "storylens.main:test"
run_with_trigger = "storylens.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
